Some functionalities added
1. parameters are flashed with a different app. Then, the main code is cleaned
to have the factory configuration. However, it could be more interesting to use
the functionality provided in the framework.

2. Some EPS_LOGI have been modified to use proper parameters.

3. it has been tried in a non success manner the internal temperature. I think
that the code must be compiled under specific version of ESP32, such as ESP32S2.
It is needed to check the exact version of the ESP to configure properly.
currently the code are commented. In the future could be added a compilation
condition.

4. Code has not been checked with unitary tests